ref(ev_msg): move parsing to separate module
